Add files via upload
Updated to create a page that will show stepwise refinement.

I was able to name the version titles but I could not figure out the syntax to create text within the boxes as your example shows "Input Buttons, Decimal Binary, Binary to Segments... etc" within each box. Unclear if this is part of the assignment as it is not shown in the provided video.